+++ lib/features/home/logic/shelves_provider.dart
@@ -85,7 +85,8 @@ Future<List<Shelf>> sortedShelves(Ref ref) async {
 
 @riverpod
 Future<List<Shelf>> rawShelves(Ref ref) async {
-  final isConnected = ref.watchConnection();
+  ref.invalidateOnReconnect();
+  final isConnected = ref.readConnection();
   if (!isConnected) {
     final downloads = await ref.read(downloadsProvider.future);
     final cache = ref.read(itemsCacheProvider.notifier);