+++ lib/features/library/logic/library_items_provider.dart
@@ -39,7 +39,7 @@ Future<List<LibraryItem>> libraryItems(Ref ref) async {
 
 @riverpod
 Future<List<LibraryItem>> rawLibraryItems(Ref ref) async {
-  ref.watchConnection();
+  ref.invalidateOnReconnect();
   final libraryId = (await ref.watch(activeLibraryDetailsProvider.future))
       .library
       .id;