+++ lib/features/library/logic/library_items_provider.g.dart
@@ -89,4 +89,4 @@ final class RawLibraryItemsProvider
   }
 }
 
-String _$rawLibraryItemsHash() => r'9737a530d00065d0e25bbd465a8997de5cbcfd6f';
+String _$rawLibraryItemsHash() => r'b3976dea1b20d02fa01af3e5c97ffad5b33544b7';